input any file

+++ Demo/pipeline.c
@@ -45,7 +45,7 @@ int main (void)
 
     int c_size = 8;
     int num_chunks = (dim_y) / (c_size);
-    double chunks[num_chunks][8][8];
+    float chunks[num_chunks][8][8];
     float result_blks[num_chunks][8][8];
     
     for(int j = 0; j < dim_y; j++){
